requirements.txt
aenum==3.1.15
aioboto3==13.4.0
aiobotocore==2.18.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioitertools==0.12.0
aiosignal==1.3.2
alembic==1.14.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==4.0.3
asyncer==0.0.8
attrs==24.3.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.36.1
botocore==1.36.1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.1
coverage==7.6.9
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.14.4
debugpy==1.8.12
decorator==5.2.1
Deprecated==1.2.15
dill==0.3.7
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
dspy==2.5.41
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.16.1
filetype==1.2.0
FLAML==2.1.2
Flask==3.0.3
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-genai==1.10.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
gradio==5.9.1
gradio_client==1.5.2
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
grpcio-tools==1.68.1
gspread==6.1.4
h11==0.14.0
h2==4.1.0
hnswlib==0.8.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.33.0
itsdangerous==2.2.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
json_repair==0.35.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langchain==0.3.13
langchain-community==0.3.1
langchain-core==0.3.28
langchain-text-splitters==0.3.4
langsmith==0.1.147
line-bot-sdk==3.14.2
litellm==1.51.0
llama-cloud==0.1.6
llama-index==0.12.7
llama-index-agent-openai==0.4.1
llama-index-cli==0.4.0
llama-index-core==0.12.7
llama-index-embeddings-jinaai==0.4.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-llms-gemini==0.4.2
llama-index-llms-openai==0.3.12
llama-index-multi-modal-llms-openai==0.4.1
llama-index-postprocessor-rankgpt-rerank==0.3.0
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.1
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.18
magicattr==0.1.6
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.1.29
monotonic==1.6
multidict==6.1.0
multiprocess==0.70.15
mypy-extensions==1.0.0
nano-vectordb==0.0.4.3
narwhals==1.19.0
neo4j==5.24.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.58.1
optuna==4.2.0
orjson==3.10.12
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
posthog==3.7.4
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
pydantic-settings==2.7.0
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pymongo==4.10.1
pyparsing==3.2.0
pypdf==5.1.0
pytest==8.3.3
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.1
qdrant-client==1.12.1
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruff==0.8.4
s3transfer==0.11.2
safehttpx==0.1.6
seaborn==0.13.2
semantic-version==2.10.0
setuptools==75.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
streamlit==1.38.0
striprtf==0.0.26
tenacity==8.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
watchdog==4.0.2
wcwidth==0.2.13
websockets==14.1
Werkzeug==3.1.3
wheel==0.44.0
wrapt==1.17.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
jasonlines